Documents convertion methods, moves to own file
